Reference for RAGTools

PromptingTools.Experimental.RAGTools.JudgeAllScores

PromptingTools.Experimental.RAGTools.JudgeRating

PromptingTools.Experimental.RAGTools.MultiIndex

PromptingTools.Experimental.RAGTools.RAGContext

PromptingTools.Experimental.RAGTools.airag

PromptingTools.Experimental.RAGTools.build_context

PromptingTools.Experimental.RAGTools.build_index

PromptingTools.Experimental.RAGTools.build_index

PromptingTools.Experimental.RAGTools.build_qa_evals

PromptingTools.Experimental.RAGTools.build_tags

PromptingTools.Experimental.RAGTools.find_closest

PromptingTools.Experimental.RAGTools.metadata_extract

PromptingTools.Experimental.RAGTools.run_qa_evals

PromptingTools.Experimental.RAGTools.run_qa_evals

PromptingTools.Experimental.RAGTools.score_retrieval_hit

PromptingTools.Experimental.RAGTools.score_retrieval_rank

PromptingTools.Experimental.RAGTools — Module

RAGTools

Provides Retrieval-Augmented Generation (RAG) functionality.

Requires: LinearAlgebra, SparseArrays, PromptingTools for proper functionality.

This module is experimental and may change at any time. It is intended to be moved to a separate package in the future.

PromptingTools.Experimental.RAGTools.JudgeAllScores — Type

final_rating is the average of all scoring criteria. Explain the final_rating in rationale

PromptingTools.Experimental.RAGTools.JudgeRating — Type

Provide the final_rating between 1-5. Provide the rationale for it.

PromptingTools.Experimental.RAGTools.MultiIndex — Type

Composite index that stores multiple ChunkIndex objects and their embeddings

PromptingTools.Experimental.RAGTools.RAGContext — Type

RAGContext

A struct for debugging RAG answers. It contains the question, answer, context, and the candidate chunks at each step of the RAG pipeline.

PromptingTools.Experimental.RAGTools.airag — Function

airag(index::AbstractChunkIndex, rag_template::Symbol = :RAGAnswerFromContext;
    question::AbstractString,
    top_k::Int = 3, `minimum_similarity::AbstractFloat`= -1.0,
    tag_filter::Union{Symbol, Vector{String}, Regex, Nothing} = :auto,
    rerank_strategy::RerankingStrategy = Passthrough(),
    model_embedding::String = PT.MODEL_EMBEDDING, model_chat::String = PT.MODEL_CHAT,
    model_metadata::String = PT.MODEL_CHAT,
    metadata_template::Symbol = :RAGExtractMetadataShort,
    chunks_window_margin::Tuple{Int, Int} = (1, 1),
    return_context::Bool = false, verbose::Bool = true,
    api_kwargs::NamedTuple = NamedTuple(),
    kwargs...)

Generates a response for a given question using a Retrieval-Augmented Generation (RAG) approach.

The function selects relevant chunks from an ChunkIndex, optionally filters them based on metadata tags, reranks them, and then uses these chunks to construct a context for generating a response.

Arguments

index::AbstractChunkIndex: The chunk index to search for relevant text.

rag_template::Symbol: Template for the RAG model, defaults to :RAGAnswerFromContext.

question::AbstractString: The question to be answered.

top_k::Int: Number of top candidates to retrieve based on embedding similarity.

minimum_similarity::AbstractFloat: Minimum similarity threshold (between -1 and 1) for filtering chunks based on embedding similarity. Defaults to -1.0.

tag_filter::Union{Symbol, Vector{String}, Regex}: Mechanism for filtering chunks based on tags (either automatically detected, specific tags, or a regex pattern). Disabled by setting to nothing.

rerank_strategy::RerankingStrategy: Strategy for reranking the retrieved chunks.

model_embedding::String: Model used for embedding the question, default is PT.MODEL_EMBEDDING.

model_chat::String: Model used for generating the final response, default is PT.MODEL_CHAT.

model_metadata::String: Model used for extracting metadata, default is PT.MODEL_CHAT.

metadata_template::Symbol: Template for the metadata extraction process from the question, defaults to: :RAGExtractMetadataShort

chunks_window_margin::Tuple{Int,Int}: The window size around each chunk to consider for context building. See ?build_context for more information.

return_context::Bool: If true, returns the context used for RAG along with the response.

verbose::Bool: If true, enables verbose logging.

api_kwargs: API parameters that will be forwarded to the API calls

Returns

If return_context is false, returns the generated message (msg).

If return_context is true, returns a tuple of the generated message (msg) and the RAG context (rag_context).

Notes

The function first finds the closest chunks to the question embedding, then optionally filters these based on tags. After that, it reranks the candidates and builds a context for the RAG model.

The tag_filter can be used to refine the search. If set to :auto, it attempts to automatically determine relevant tags (if index has them available).

The chunks_window_margin allows including surrounding chunks for richer context, considering they are from the same source.

The function currently supports only single ChunkIndex.

Examples

Using airag to get a response for a question:

index = build_index(...)  # create an index
question = "How to make a barplot in Makie.jl?"
msg = airag(index, :RAGAnswerFromContext; question)

# or simply
msg = airag(index; question)

See also build_index, build_context, CandidateChunks, find_closest, find_tags, rerank

PromptingTools.Experimental.RAGTools.build_context — Method

build_context(index::AbstractChunkIndex, reranked_candidates::CandidateChunks; chunks_window_margin::Tuple{Int, Int}) -> Vector{String}

Build context strings for each position in reranked_candidates considering a window margin around each position.

Arguments

reranked_candidates::CandidateChunks: Candidate chunks which contain positions to extract context from.

index::ChunkIndex: The index containing chunks and sources.

chunks_window_margin::Tuple{Int, Int}: A tuple indicating the margin (before, after) around each position to include in the context. Defaults to (1,1), which means 1 preceding and 1 suceeding chunk will be included. With (0,0), only the matching chunks will be included.

Returns

Vector{String}: A vector of context strings, each corresponding to a position in reranked_candidates.

Examples

index = ChunkIndex(...)  # Assuming a proper index is defined
candidates = CandidateChunks(index.id, [2, 4], [0.1, 0.2])
context = build_context(index, candidates; chunks_window_margin=(0, 1)) # include only one following chunk for each matching chunk

PromptingTools.Experimental.RAGTools.build_index — Function

Build an index for RAG (Retriever-Augmented Generation) applications. REQUIRES SparseArrays and LinearAlgebra packages to be loaded!!

PromptingTools.Experimental.RAGTools.build_index — Method

build_index(files::Vector{<:AbstractString};
    separators = ["

", ". ", " "], maxlength::Int = 256, extractmetadata::Bool = false, verbose::Bool = true, metadatatemplate::Symbol = :RAGExtractMetadataShort, modelembedding::String = PT.MODELEMBEDDING, modelmetadata::String = PT.MODELCHAT, apikwargs::NamedTuple = NamedTuple())

Build an index for RAG (Retriever-Augmented Generation) applications from the provided file paths. The function processes each file, splits its content into chunks, embeds these chunks, optionally extracts metadata, and then compiles this information into a retrievable index.

Arguments

files: A vector of valid file paths to be indexed.

separators: A list of strings used as separators for splitting the text in each file into chunks. Default is `["

", ". ", " "]`.

max_length: The maximum length of each chunk (if possible with provided separators). Default is 256.

extract_metadata: A boolean flag indicating whether to extract metadata from each chunk (to build filter tags in the index). Default is false. Metadata extraction incurs additional cost and requires model_metadata and metadata_template to be provided.

verbose: A boolean flag for verbose output. Default is true.

metadata_template: A symbol indicating the template to be used for metadata extraction. Default is :RAGExtractMetadataShort.

model_embedding: The model to use for embedding.

model_metadata: The model to use for metadata extraction.

api_kwargs: Parameters to be provided to the API endpoint.

Returns

ChunkIndex: An object containing the compiled index of chunks, embeddings, tags, vocabulary, and sources.

See also: MultiIndex, CandidateChunks, find_closest, find_tags, rerank, airag

Examples

# Assuming `test_files` is a vector of file paths
index = build_index(test_files; max_length=10, extract_metadata=true)

# Another example with metadata extraction and verbose output
index = build_index(["file1.txt", "file2.txt"]; 
                    separators=[". "], 
                    extract_metadata=true, 
                    verbose=true)

PromptingTools.Experimental.RAGTools.build_qa_evals — Method

build_qa_evals(doc_chunks::Vector{<:AbstractString}, sources::Vector{<:AbstractString};
               model=PT.MODEL_CHAT, instructions="None.", qa_template::Symbol=:RAGCreateQAFromContext, 
               verbose::Bool=true, api_kwargs::NamedTuple = NamedTuple(), kwargs...) -> Vector{QAEvalItem}

Create a collection of question and answer evaluations (QAEvalItem) from document chunks and sources. This function generates Q&A pairs based on the provided document chunks, using a specified AI model and template.

Arguments

doc_chunks::Vector{<:AbstractString}: A vector of document chunks, each representing a segment of text.

sources::Vector{<:AbstractString}: A vector of source identifiers corresponding to each chunk in doc_chunks (eg, filenames or paths).

model: The AI model used for generating Q&A pairs. Default is PT.MODEL_CHAT.

instructions::String: Additional instructions or context to provide to the model generating QA sets. Defaults to "None.".

qa_template::Symbol: A template symbol that dictates the AITemplate that will be used. It must have placeholder context. Default is :CreateQAFromContext.

api_kwargs::NamedTuple: Parameters that will be forwarded to the API endpoint.

verbose::Bool: If true, additional information like costs will be logged. Defaults to true.

Returns

Vector{QAEvalItem}: A vector of QAEvalItem structs, each containing a source, context, question, and answer. Invalid or empty items are filtered out.

Notes

The function internally uses aiextract to generate Q&A pairs based on the provided qa_template. So you can use any kwargs that you want.

Each QAEvalItem includes the context (document chunk), the generated question and answer, and the source.

The function tracks and reports the cost of AI calls if verbose is enabled.

Items where the question, answer, or context is empty are considered invalid and are filtered out.

Examples

Creating Q&A evaluations from a set of document chunks:

doc_chunks = ["Text from document 1", "Text from document 2"]
sources = ["source1", "source2"]
qa_evals = build_qa_evals(doc_chunks, sources)

PromptingTools.Experimental.RAGTools.build_tags — Function

Builds a matrix of tags and a vocabulary list. REQUIRES SparseArrays and LinearAlgebra packages to be loaded!!

PromptingTools.Experimental.RAGTools.find_closest — Method

find_closest(emb::AbstractMatrix{<:Real},
    query_emb::AbstractVector{<:Real};
    top_k::Int = 100, minimum_similarity::AbstractFloat = -1.0)

Finds the indices of chunks (represented by embeddings in emb) that are closest (cosine similarity) to query embedding (query_emb).

If minimum_similarity is provided, only indices with similarity greater than or equal to it are returned. Similarity can be between -1 and 1 (-1 = completely opposite, 1 = exactly the same).

Returns only top_k closest indices.

PromptingTools.Experimental.RAGTools.metadata_extract — Method

metadata_extract(item::MetadataItem)
metadata_extract(items::Vector{MetadataItem})

Extracts the metadata item into a string of the form category:::value (lowercased and spaces replaced with underscores).

Example

msg = aiextract(:RAGExtractMetadataShort; return_type=MaybeMetadataItems, text="I like package DataFrames", instructions="None.")
metadata = metadata_extract(msg.content.items)

PromptingTools.Experimental.RAGTools.run_qa_evals — Method

run_qa_evals(index::AbstractChunkIndex, qa_items::AbstractVector{<:QAEvalItem};
    api_kwargs::NamedTuple = NamedTuple(),
    airag_kwargs::NamedTuple = NamedTuple(),
    qa_evals_kwargs::NamedTuple = NamedTuple(),
    verbose::Bool = true, parameters_dict::Dict{Symbol, <:Any} = Dict{Symbol, Any}())

Evaluates a vector of QAEvalItems and returns a vector QAEvalResult. This function assesses the relevance and accuracy of the answers generated in a QA evaluation context.

See ?run_qa_evals for more details.

Arguments

qa_items::AbstractVector{<:QAEvalItem}: The vector of QA evaluation items containing the questions and their answers.

verbose::Bool: If true, enables verbose logging. Defaults to true.

api_kwargs::NamedTuple: Parameters that will be forwarded to the API calls. See ?aiextract for details.

airag_kwargs::NamedTuple: Parameters that will be forwarded to airag calls. See ?airag for details.

qa_evals_kwargs::NamedTuple: Parameters that will be forwarded to run_qa_evals calls. See ?run_qa_evals for details.

parameters_dict::Dict{Symbol, Any}: Track any parameters used for later evaluations. Keys must be Symbols.

Returns

Vector{QAEvalResult}: Vector of evaluation results that includes various scores and metadata related to the QA evaluation.

Example

index = "..." # Assuming a proper index is defined
qa_items = [QAEvalItem(question="What is the capital of France?", answer="Paris", context="France is a country in Europe."),
            QAEvalItem(question="What is the capital of Germany?", answer="Berlin", context="Germany is a country in Europe.")]

# Let's run a test with `top_k=5`
results = run_qa_evals(index, qa_items; airag_kwargs=(;top_k=5), parameters_dict=Dict(:top_k => 5))

# Filter out the "failed" calls
results = filter(x->!isnothing(x.answer_score), results);

# See average judge score
mean(x->x.answer_score, results)

PromptingTools.Experimental.RAGTools.run_qa_evals — Method

run_qa_evals(qa_item::QAEvalItem, ctx::RAGContext; verbose::Bool = true,
             parameters_dict::Dict{Symbol, <:Any}, judge_template::Symbol = :RAGJudgeAnswerFromContext,
             model_judge::AbstractString, api_kwargs::NamedTuple = NamedTuple()) -> QAEvalResult

Evaluates a single QAEvalItem using a RAG context (RAGContext) and returns a QAEvalResult structure. This function assesses the relevance and accuracy of the answers generated in a QA evaluation context.

Arguments

qa_item::QAEvalItem: The QA evaluation item containing the question and its answer.

ctx::RAGContext: The context used for generating the QA pair, including the original context and the answers. Comes from airag(...; return_context=true)

verbose::Bool: If true, enables verbose logging. Defaults to true.

parameters_dict::Dict{Symbol, Any}: Track any parameters used for later evaluations. Keys must be Symbols.

judge_template::Symbol: The template symbol for the AI model used to judge the answer. Defaults to :RAGJudgeAnswerFromContext.

model_judge::AbstractString: The AI model used for judging the answer's quality. Defaults to standard chat model, but it is advisable to use more powerful model GPT-4.

api_kwargs::NamedTuple: Parameters that will be forwarded to the API endpoint.

Returns

QAEvalResult: An evaluation result that includes various scores and metadata related to the QA evaluation.

Notes

The function computes a retrieval score and rank based on how well the context matches the QA context.

It then uses the judge_template and model_judge to score the answer's accuracy and relevance.

In case of errors during evaluation, the function logs a warning (if verbose is true) and the answer_score will be set to nothing.

Examples

Evaluating a QA pair using a specific context and model:

qa_item = QAEvalItem(question="What is the capital of France?", answer="Paris", context="France is a country in Europe.")
ctx = RAGContext(source="Wikipedia", context="France is a country in Europe.", answer="Paris")
parameters_dict = Dict("param1" => "value1", "param2" => "value2")

eval_result = run_qa_evals(qa_item, ctx, parameters_dict=parameters_dict, model_judge="MyAIJudgeModel")

PromptingTools.Experimental.RAGTools.score_retrieval_hit — Method

Returns 1.0 if context overlaps or is contained within any of the candidate_context

PromptingTools.Experimental.RAGTools.score_retrieval_rank — Method

Returns Integer rank of the position where context overlaps or is contained within a candidate_context